fixing norm
