+++ src/input_split/input_list_join.c
@@ -1,0 +1,67 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   input_list_join.c                                  :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: paulhenr <paulhenr@student.42.rio>         +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/03/27 10:28:51 by paulhenr          #+#    #+#             */
+/*   Updated: 2024/03/27 10:29:34 by paulhenr         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "input_handler.h"
+
+char	*ft_strjoin_token_lst(t_list2 *list)
+{
+	char	*joined;
+	size_t	len;
+	t_list2	*tmp;
+
+	len = 0;
+	tmp = list;
+	if (!list)
+		return (ft_perror(__func__, ARGNULL), NULL);
+	while (tmp)
+	{
+		len += ft_strlen(((t_token *)tmp->data)->value);
+		tmp = tmp->next;
+	}
+	joined = malloc((len + 1) * sizeof(char));
+	*joined = '\0';
+	if (!joined)
+		return (perror(__func__), NULL);
+	while (list)
+	{
+		ft_strcat(joined, ((t_token *)list->data)->value);
+		list = list->next;
+	}
+	return (joined);
+}
+
+char	*ft_strjoinlst(t_list2 *list)
+{
+	char	*joined;
+	size_t	len;
+	t_list2	*tmp;
+
+	len = 0;
+	tmp = list;
+	if (!list)
+		return (ft_perror(__func__, ARGNULL), NULL);
+	while (tmp)
+	{
+		len += ft_strlen((char *)tmp->data);
+		tmp = tmp->next;
+	}
+	joined = malloc((len + 1) * sizeof(char));
+	*joined = '\0';
+	if (!joined)
+		return (perror(__func__), NULL);
+	while (list)
+	{
+		ft_strcat(joined, (char *)list->data);
+		list = list->next;
+	}
+	return (joined);
+}
