+++ src/input_split/input_split_utils.c
@@ -6,7 +6,7 @@
 /*   By: paulhenr <paulhenr@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/21 12:39:53 by paulhenr          #+#    #+#             */
-/*   Updated: 2024/03/26 15:06:32 by paulhenr         ###   ########.fr       */
+/*   Updated: 2024/03/27 10:29:09 by paulhenr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -83,31 +83,4 @@ static int	classify_arg(t_list2 *list)
 			return (HERE_DOC_ARG);
 	}
 	return (ARG);
-}
-
-char	*ft_strjoin_token_lst(t_list2 *list)
-{
-	char	*joined;
-	size_t	len;
-	t_list2	*tmp;
-
-	len = 0;
-	tmp = list;
-	if (!list)
-		return (ft_perror(__func__, ARGNULL), NULL);
-	while (tmp)
-	{
-		len += ft_strlen(((t_token *)tmp->data)->value);
-		tmp = tmp->next;
-	}
-	joined = malloc((len + 1) * sizeof(char));
-	*joined = '\0';
-	if (!joined)
-		return (perror(__func__), NULL);
-	while (list)
-	{
-		ft_strcat(joined, ((t_token *)list->data)->value);
-		list = list->next;
-	}
-	return (joined);
 }
